divided index.ts into server.ts and index.ts
